requirements.txt
Flask
gunicorn
pandas
openpyxl
psycopg2-binary
python-dotenv